+++ app/models/topic.rb
@@ -32,7 +32,12 @@ class Topic < ApplicationRecord
   after_create_commit :create_first_reply
 
   # Scopes
-  scope :shadow_banned, ->(arg) { joins(:user).merge(User.where(shadow_banned: [nil, false]).or(User.where(id: arg))) }
+  scope :shadow_banned, ->(arg) do
+    joins(:user).merge(
+      User.where(shadow_banned: [nil, false])
+      .or(User.where(id: arg))
+    )
+  end
   # Auto-lock after replies have ceased for some time.
   scope :not_auto_locked, -> { where(locked: false).where("DATE(last_reply_at) >= ?", Time.zone.today - AUTO_LOCK_IN) }
   scope :auto_locked, -> { where(locked: true).or(where("DATE(last_reply_at) < ?", Time.zone.today - AUTO_LOCK_IN)) }